Support ignoring assembly resolution failures
Added a new command line argument -ignore-unresolved-assemblies
or -iua which is false by default. This can be set so that
if an assembly cannot be resolved an exception is not thrown and
instead a warning is logged. For example, an assembly may
referenced MonoAndroid but you may not be interested in the
intersection with any Android types so instead of having the
ApiIntersect tool fail when it fails to resolve MonoAndroid you
can optionally have it ignore this resolution failure and
continue processing.